moved fixtures into conftest.py
